feat: Add cli search for v2
